requirements.txt
fastapi==0.106.0
uvicorn[standard]==0.22.0
pydantic
aioredis==2.0.0
google-cloud-bigquery==3.8.0